+++ lib/libs/shared/src/types/payment-service/ibank-account.type.d.ts
@@ -2,8 +2,8 @@ import { tags } from 'typia';
 import { BaseResponse } from '../common.type';
 export declare namespace IBankAccount {
     interface ICreateBankAccount {
-        shop_id?: number & tags.Type<'uint32'>;
-        qr_image_id?: number & tags.Type<'uint32'>;
+        shop_id?: number & tags.Type<'uint64'>;
+        qr_image_id?: number & tags.Type<'uint64'>;
         bankName: string;
         accountName: string;
         accountNumber: string;
@@ -12,7 +12,7 @@ export declare namespace IBankAccount {
         active?: boolean;
     }
     interface IUpdateBankAccount {
-        qr_image_id?: number & tags.Type<'uint32'>;
+        qr_image_id?: number & tags.Type<'uint64'>;
         bankName?: string;
         accountName?: string;
         accountNumber?: string;
