+++ lib/libs/shared/src/types/payment-service/icheckout.type.d.ts
@@ -2,9 +2,9 @@ import { tags } from 'typia';
 import { BaseResponse } from '../common.type';
 export declare namespace ICheckout {
     interface ICreateCheckout {
-        shop_id?: number & tags.Type<'uint32'>;
-        partner_id: (number & tags.Type<'uint32'>) | (string & tags.Format<'uuid'>);
-        order_id?: number & tags.Type<'uint32'>;
+        shop_id?: number & tags.Type<'uint64'>;
+        partner_id: (number & tags.Type<'uint64'>) | (string & tags.Format<'uuid'>);
+        order_id?: number & tags.Type<'uint64'>;
         uuid?: string;
         currency?: string;
         presentmentCurrency?: string;
