+++ lib/libs/shared/src/types/product-service/icategory.d.ts
@@ -3,8 +3,8 @@ import { BaseResponse } from '../common.type';
 export declare namespace ICategory {
     interface IReqCreateCategory {
         name: string;
-        parent_id?: number & tags.Type<'uint32'>;
-        image_id?: number & tags.Type<'uint32'>;
+        parent_id?: number & tags.Type<'uint64'>;
+        image_id?: number & tags.Type<'uint64'>;
         icon?: string;
         active?: boolean;
         description?: string;
@@ -12,13 +12,13 @@ export declare namespace ICategory {
         userId?: string;
     }
     interface ICreateCategory extends IReqCreateCategory {
-        shop_id: number & tags.Type<'uint32'>;
+        shop_id: number & tags.Type<'uint64'>;
     }
     interface IUpdateCategory {
         name?: string;
         active?: boolean;
         icon?: string | null | undefined;
-        image_id?: number & tags.Type<'uint32'>;
+        image_id?: number & tags.Type<'uint64'>;
         description?: string | null;
         htmlContent?: string | null;
         userId?: string | null;
