+++ lib/libs/shared/src/types/product-service/iselection-group.d.ts
@@ -3,18 +3,18 @@ import { BaseResponse } from '../common.type';
 import { ISelection } from './iselection';
 export declare namespace ISelectionGroup {
     interface IReqCreateSelectionGroup {
-        image_id?: number & tags.Type<'uint32'>;
+        image_id?: number & tags.Type<'uint64'>;
         name: string;
         active?: boolean;
         icon?: string | null;
         userId?: string;
     }
     interface ICreateSelectionGroup extends IReqCreateSelectionGroup {
-        shop_id: number & tags.Type<'uint32'>;
+        shop_id: number & tags.Type<'uint64'>;
     }
     interface IUpdateSelectionGroup {
         id?: string;
-        image_id?: number & tags.Type<'uint32'>;
+        image_id?: number & tags.Type<'uint64'>;
         name?: string;
         active?: boolean;
         icon?: string | null | undefined;
